Improve object handling
In order to allow more complex scene rendering in the future, a host of
new classes have been added that represent a scene and the objects
within that scene.

Not only does this significantly clean up code related to objects, it
also allows multiple objects in a scene to share a single model,
reducing GPU memory footprint for future cases where multiple of the
same model may exist at once.

Additionally, a new shader has been added that renders the normal
vectors as colours. This shader is intended for debugging purposes to
make it easier to visualize the normal vectors of an object.