Implement Phong shading/illumination
The previous debug shader which displayed normal vectors as colours has
now been replaced by a full implementation of Phong shading and
illumination.

The Phong fragment shader has a hardcoded maximum of 16 point lights
that can exist at one time. Currently, one light is hardcoded, but this
will likely be changed in the future to allow light information to be
read from a file.