Allow enabling/disabling lighting
For debugging purposes, it is often useful to disable lighting and see
the scene in full brightness. This has now been implemented and lighting
can be toggled on and off by pressing the L key.

Disabling lighting is accomplished by setting the scene ambient lighting
to full white light and not sending any lights. This has the effect of
disabling lighting calculations without requiring a different shader
program.