Set far plane to infinity
Previously, the far plane was being set to an arbitrary number. However,
this does not meet the assignment requirements since it is possible for
the object to be clipped beyond the far plane.

To prevent this and make everything more pleasant in general, the far
plane has been set to infinity. This guarantees that nothing will ever
clip beyond the far plane without having to perform any calculation of
what the far plane should be.