Fix camera coordinates becoming (NaN, NaN, NaN)
The arcball rotation handler in OrbitControls had a nasty bug in it that
could sometimes cause the camera coordinates and orientation vector to
become full of NaNs. Once this happened, nothing would display on the
viewport and the NaNs would never go away.

This was caused when the arcball projection of the user's cursor
position was the same as the arcball projection of their previous cursor
position. When this happened, their cross product was the null vector,
meaning that the program would attempt to rotate the camera about the
null vector, which caused NaNs to appear. This has been corrected so
that when the axis of rotation is the null vector, the rotation is
cancelled prematurely.

Additionally, pressing the C key will now reset the camera position and
orientation for convenience.