Implement object transformations
As per the assignment description, it is necessary to implement certain
modeling transformations on the loaded object; the existing camera
movements were not enough to satisfy this requirement. The required
transformations are: movement, intrinsic rotation (yaw, pitch, and
roll), and scaling.

The loaded object can be moved around the scene using WASDQE while not
holding shift. Rotation can be done by using WASDQE while holding shift.
Scaling can be done using Z and X.

The speeds at which these transforms are applied can be adjusted by
using the I and K keys to increase and decrease the transformation speed
respectively.