Implement scene file loading
In order to allow scenes with multiple objects, as will be required to
model the chessboard for the bonus part of the assignment, it is
necessary to have more complicated and controllable logic to load the
world.

A new scene file format has been added which allows this to be done.
This file contains paths to the necessary OBJ models to load,
definitions for the materials used in the scene, as well as all
configuration information for all present lights and objects. A file of
this type is now expected and will be parsed to populate the world.