Allow editing all objects in a scene
Previously, the editing controls only allowed the user to perform
transformations on the first loaded object. However, this was not
versatile enough to edit scenes properly. Thus, it is now possible to
use tab and shift-tab to select an object to edit in the scene.

Additionally, pressing the P key will print position, rotation, and
scaling information to standard output in the same format used by scene
files. This is to allow transformations to be edited and then saved into
a scene file without needing to use an external program.