Set texture sample modes properly
Previously, the texture sample modes for ambient occlusion, diffuse, and
specular maps were not being set correctly, causing the textures to look
very bad from a distance. All of these textures will now be properly set
to linear interpolation.

It would theoretically be possible to use mipmaps for downscaling, but
the unwrapped texture maps are not created with this in mind; as a
result of this, trying to use mipmaps would cause UV seams to become
very apparent.