Add some new scenes
Two new scenes have been added: a scene containing only a coffee cup and
a scene containing a chess board on a table (set up at the ending
position from Mikhail Botvinnik vs Mikhail M Yudovich Sr. from 1933).

The coffee cup scene is meant to better show off specular reflection, as
the coffee cup is set to be have a high shininess factor.

The chess board scene is for the bonus part of the assignment. It
contains all 32 chess pieces arranged on a chess board on top of an oak
table with some oak chairs. It also contains 2 different point lights
for better ambiance.