Implement scene ambient lighting
Previously, there was no way to specify the scene ambient lighting in a
scene file other than to create a point light with only an ambient
component and no attenuation. The phong fragment shader did have the
ability to accept a scene ambient light component, but this was only
used to enable full brightness mode.

The "alight" command has now been added to scene files, which allows
this value to be controlled.