Prevent material redefinition
Previously, it was possible to define a material with the same name as a
previously defined material, overwriting it only for future uses of the
given material name. This was despite the fact that object models could
not be redefined.

This has been corrected and redefinitions of materials are now
disallowed.